# api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: TMA Server API
  version: 2.0.0-alpha.1
  description: User and control-plane contract for the TMA Go Core SDK.
servers:
  - url: /
paths:
...
  /v2/sessions/{session_id}/runs:
    get:
      operationId: get_v2_sessions_by_session_id_runs
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
        default:
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: post_v2_sessions_by_session_id_runs
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRunRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartRunResponse"
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartRunResponse"
        default:
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v2/sessions/{session_id}/runs/{run_id}:
    get:
      operationId: get_v2_sessions_by_session_id_runs_by_run_id
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
        - name: run_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v2/sessions/{session_id}/runs/{run_id}/cancel:
    post:
      operationId: post_v2_sessions_by_session_id_runs_by_run_id_cancel
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
        - name: run_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v2/sessions/{session_id}/runs/{run_id}/events:
    get:
      operationId: get_v2_sessions_by_session_id_runs_by_run_id_events
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
        - name: run_id
          in: path
          required: true
          schema: {type: string}
        - name: after_seq
          in: query
          required: false
          schema: {type: string}
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
        default:
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v2/sessions/{session_id}/runs/{run_id}/events/stream:
    get:
      operationId: get_v2_sessions_by_session_id_runs_by_run_id_events_stream
      parameters:
        - name: session_id
          in: path
          required: true
          schema: {type: string}
        - name: run_id
          in: path
          required: true
          schema: {type: string}
        - name: after_seq
          in: query
          required: false
          schema: {type: string}
      responses:
        "200":
          description: Successful response
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/EventStream"
        default:
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /v2/workers:
    get:
      operationId: get_v2_workers
      parameters:
        - name: workspace_id
          in: query
          required: false
          schema: {type: string}
        - name: status
          in: query
          required: false
          schema: {type: string}
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerList"
        default:
          description: API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message, request_id, retryable]
      properties:
        code: {type: string, pattern: "^[a-z][a-z0-9_]*$"}
        message: {type: string}
        request_id: {type: string}
        retryable: {type: boolean}
        details:
          type: object
          additionalProperties: true
          x-tma-dynamic-json: true
          description: Stable machine-readable error details whose fields depend on error.code.
    DynamicJSONValue:
      description: Extension JSON whose shape is selected by the surrounding event, schema, provider, or runtime contract.
      x-tma-dynamic-json: true
    Event:
      type: object
      required: [id, session_id, seq, type, created_at]
      properties:
        id: {type: string}
        session_id: {type: string}
        turn_id: {type: string}
        seq: {type: integer, format: int64, maximum: 9007199254740991}
        type: {type: string}
        payload: {$ref: "#/components/schemas/DynamicJSONValue"}
        created_at: {type: string, format: date-time}
...
    EventList:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: {$ref: "#/components/schemas/Event"}
    EventStream:
      type: string
      description: Server-sent event stream whose data fields contain Event JSON.
...
    Run:
      type: object
      required: [id, session_id, status, attempt, started_at]
      properties:
        id: {type: string}
        session_id: {type: string}
        status: {type: string, enum: [running, waiting_approval, completed, failed, interrupted]}
        user_event_id: {type: string}
        user_event_seq: {type: integer, format: int64, maximum: 9007199254740991}
        attempt: {type: integer, format: int32}
        started_at: {type: string, format: date-time}
        ended_at: {type: string, format: date-time, nullable: true}
        interrupt_requested_at: {type: string, format: date-time, nullable: true}
        error_message: {type: string}
        idempotency_key: {type: string}
    RunList:
      type: object
      required: [runs]
      properties:
        runs: {type: array, items: {$ref: "#/components/schemas/Run"}}
    StartRunRequest:
      type: object
      required: [input]
      properties:
        input: {type: object, additionalProperties: true, x-tma-dynamic-json: true}
        idempotency_key: {type: string}
    StartRunResponse:
      type: object
      required: [run, created]
      properties:
        run: {$ref: "#/components/schemas/Run"}
        events: {type: array, items: {$ref: "#/components/schemas/Event"}}
        created: {type: boolean}
...
    Worker:
      type: object
      required: [id, workspace_id, name, worker_type, status, registered_by, registered_at]
      properties:
        id: {type: string}
        workspace_id: {type: string}
        name: {type: string}
        worker_type: {type: string}
        status: {type: string}
        capabilities: {type: object, additionalProperties: true, x-tma-dynamic-json: true}
        metadata: {type: object, additionalProperties: true, x-tma-dynamic-json: true}
        registered_by: {type: string}
        registered_at: {type: string, format: date-time}
        last_seen_at: {type: string, format: date-time, nullable: true}
        lease_expires_at: {type: string, format: date-time, nullable: true}
        archived_at: {type: string, format: date-time, nullable: true}
    WorkerList:
      type: object
      required: [workers]
      properties:
        workers: {type: array, items: {$ref: "#/components/schemas/Worker"}}
...